+++ packages/gg-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @kenkaiiii/gg-core
 
+## 5.31.0
+
+### Patch Changes
+
+- @kenkaiiii/gg-ai@5.31.0
+
 ## 5.30.3
 
 ### Patch Changes